\name{EWOC2-internal}
\alias{Function.decimalplaces}
\alias{Function.fdist}
\alias{Function.generate.bugs.file}
\alias{Function.Mi.Dose.Increment_simu}
\alias{Function.mtd_logistic}
\alias{Function.MTDdoses}
\alias{Function.nextdose.2d}
\alias{Function.parameter}
\alias{Function.postdlt}
\alias{Function.simu.2d}

\title{Internal functions for EWOC2}
\description{Internal functions for EWOC2}
\details{These are not to be called by the user.}
\keyword{internal}