\name{EWOC2-package}
\alias{EWOC2-package}
\alias{EWOC2}
\docType{package}
\title{
\packageTitle{EWOC2}
}
\description{
\packageDescription{EWOC2}
}
\author{
\packageAuthor{EWOC2}

Maintainer: \packageMaintainer{EWOC2}
}
\references{Tighiouart M, Li Q and Rogatko A. A Bayesian adaptive design for estimating the maximuym tolerated dose curve using drug combinations in cancer phase I clinical trials. Statistics in Medicine. 2017, 36: 280-290.}
\keyword{ package }
